Keep the existing Lua table format
For the data page, use the Lua table format already in use on the
English Wikipedia. This includes a top comment explaining the purpose of
the page, and newlines between different templates. This change gets the
top comment by parsing it from the current data page, meaning changes to
it added by other editors will be persisted.