Bug fix: redirects to different pages are included in Lua table
Fix a bug where redirects to different are sometimes included in the Lua
table. This happens when the page in question includes a normal link to
the page being queried, and is also a redirect targeting a different
page. This occurs because fetch_template_metadata uses the backlinks()
method instead of the redirects() method. The backlinks method uses the
backlinks generator in the query API, which returns all links to the
page, not just redirects that have the page as the target.

Fixes #1